*9003* Introduce new hooks TemplateManager::fetch and FileManager::deleteFile

+++ classes/file/FileManager.inc.php
@@ -289,10 +289,11 @@ class FileManager {
 	 */
 	function deleteFile($filePath) {
 		if ($this->fileExists($filePath)) {
+			$result = null;
+			if (HookRegistry::call('FileManager::deleteFile', array($filePath, &$result))) return $result;
 			return unlink($filePath);
-		} else {
-			return false;
 		}
+		return false;
 	}
 
 	/**
@@ -406,7 +407,7 @@ class FileManager {
 				return DOCUMENT_TYPE_ZIP;
 			case 'application/epub':
 			case 'application/epub+zip':
-				return DOCUMENT_TYPE_EPUB;	
+				return DOCUMENT_TYPE_EPUB;
 			default:
 				return DOCUMENT_TYPE_DEFAULT;
 		}
